# server/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityOut": {
        "properties": {
          "activity_type": {
            "enum": [
              "running",
              "cycling",
              "walking"
            ],
            "title": "Activity Type",
            "type": "string"
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisOut"
          },
          "client_activity_id": {
            "title": "Client Activity Id",
            "type": "string"
          },
          "client_summary": {
            "additionalProperties": true,
            "title": "Client Summary",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "ended_at": {
            "format": "date-time",
            "title": "Ended At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "source_app_version": {
            "title": "Source App Version",
            "type": "string"
          },
          "source_platform": {
            "title": "Source Platform",
            "type": "string"
          },
          "split_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SplitPlanOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "client_activity_id",
          "activity_type",
          "started_at",
          "ended_at",
          "title",
          "notes",
          "device_name",
          "client_summary",
          "source_platform",
          "source_app_version",
          "created_at",
          "updated_at",
          "analysis",
          "tags"
        ],
        "title": "ActivityOut",
        "type": "object"
      },
      "ActivityPatchRequest": {
        "properties": {
          "device_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ActivityPatchRequest",
        "type": "object"
      },
...
      "AnalysisOut": {
        "properties": {
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "pending",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "AnalysisOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SplitPlanOut": {
        "description": "The split plan an activity was uploaded with (issue #100) \u2014 mirrors\nActivity.split_plan/app.analysis.gpx_parser.SplitPlanData. None on\nActivityOut for an activity with no plan at all (an old upload, or a\nplain rolling preference with no target).",
        "properties": {
          "custom_splits": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              ],
              "type": "array"
            },
            "title": "Custom Splits",
            "type": "array"
          },
          "rolling_target_mps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rolling Target Mps"
          },
          "targets_as": {
            "enum": [
              "pace",
              "speed"
            ],
            "title": "Targets As",
            "type": "string"
          }
        },
        "required": [
          "rolling_target_mps",
          "custom_splits",
          "targets_as"
        ],
        "title": "SplitPlanOut",
        "type": "object"
      },
...
      "TagOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Simple Activity Tracker Server",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/activities/{activity_id}": {
      "delete": {
        "operationId": "delete_activity_api_v1_activities__activity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Activity",
        "tags": [
          "activities"
        ]
      },
      "get": {
        "operationId": "get_activity_api_v1_activities__activity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Activity",
        "tags": [
          "activities"
        ]
      },
      "patch": {
        "operationId": "patch_activity_api_v1_activities__activity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Activity",
        "tags": [
          "activities"
        ]
      }
    },
...
  }
}